quiz_servise/requirements.txt
Flask==2.0.2
grpcio-tools~=1.30